Added refactoring for cchanging bidirectional association to unidirectional association
