+++ java/src/test/java/io/cucumber/cucumberexpressions/ParameterTypeRegistryTest.java
@@ -39,6 +39,8 @@ public class ParameterTypeRegistryTest {
 
     private final ParameterTypeRegistry registry = new ParameterTypeRegistry(Locale.ENGLISH);
 
+
+
     @Test
     public void does_not_allow_more_than_one_preferential_parameter_type_for_each_regexp() {
 
@@ -58,54 +60,7 @@ public class ParameterTypeRegistryTest {
         assertThat("Unexpected message", thrownException.getMessage(), is(equalTo("There can only be one preferential parameter type per regexp. The regexp /[A-Z]+\\w+/ is used for two preferential parameter types, {name} and {place}")));
     }
 
-    @Test
-    public void looks_up_preferential_parameter_type_by_regexp() {
-        ParameterType<Name> name = new ParameterType<>("name", CAPITALISED_WORD, Name.class, Name::new, false, false);
-        ParameterType<Person> person = new ParameterType<>("person", CAPITALISED_WORD, Person.class, Person::new, false, true);
-        ParameterType<Place> place = new ParameterType<>("place", CAPITALISED_WORD, Place.class, Place::new, false, false);
-        registry.defineParameterType(name);
-        registry.defineParameterType(person);
-        registry.defineParameterType(place);
-        assertSame(person, registry.lookupByRegexp(CAPITALISED_WORD, Pattern.compile("([A-Z]+\\w+) and ([A-Z]+\\w+)"), "Lisa and Bob"));
-    }
 
-    @Test
-    public void throws_ambiguous_exception_on_lookup_when_no_parameter_types_are_preferential() {
-        ParameterType<Name> name = new ParameterType<>("name", CAPITALISED_WORD, Name.class, Name::new, true, false);
-        ParameterType<Person> person = new ParameterType<>("person", CAPITALISED_WORD, Person.class, Person::new, true, false);
-        ParameterType<Place> place = new ParameterType<>("place", CAPITALISED_WORD, Place.class, Place::new, true, false);
-        registry.defineParameterType(name);
-        registry.defineParameterType(person);
-        registry.defineParameterType(place);
-
-        String expected = "" +
-                "Your Regular Expression /([A-Z]+\\w+) and ([A-Z]+\\w+)/\n" +
-                "matches multiple parameter types with regexp /[A-Z]+\\w+/:\n" +
-                "   {name}\n" +
-                "   {person}\n" +
-                "   {place}\n" +
-                "\n" +
-                "I couldn't decide which one to use. You have two options:\n" +
-                "\n" +
-                "1) Use a Cucumber Expression instead of a Regular Expression. Try one of these:\n" +
-                "   {name} and {name}\n" +
-                "   {name} and {person}\n" +
-                "   {name} and {place}\n" +
-                "   {person} and {name}\n" +
-                "   {person} and {person}\n" +
-                "   {person} and {place}\n" +
-                "   {place} and {name}\n" +
-                "   {place} and {person}\n" +
-                "   {place} and {place}\n" +
-                "\n" +
-                "2) Make one of the parameter types preferential and continue to use a Regular Expression.\n" +
-                "\n";
-
-        final Executable testMethod = () -> registry.lookupByRegexp(CAPITALISED_WORD, Pattern.compile("([A-Z]+\\w+) and ([A-Z]+\\w+)"), "Lisa and Bob");
-
-        final AmbiguousParameterTypeException thrownException = assertThrows(AmbiguousParameterTypeException.class, testMethod);
-        assertThat("Unexpected message", thrownException.getMessage(), is(equalTo(expected)));
-    }
 
     @Test
     public void does_not_allow_anonymous_parameter_type_to_be_registered() {
